requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto==2.49.0
boto3==1.34.74
botocore==1.34.74
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==42.0.5
dj-database-url==2.1.0
Django==4.2.11
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-redis==5.4.0
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
kombu==5.3.6
packaging==24.0
paypalrestsdk==1.13.3
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
redis==5.0.3
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
